+++ Assets/Script/CommonLib/Responses/VerifyStageBattleResponse.cs
@@ -6,7 +6,7 @@ namespace Script.CommonLib.Responses
     public class VerifyStageBattleResponse
     {
         public TeamFlag Winner { get; set; }
-        public List<Tuple<uint, float>> AliveEntities { get; set; }
+        public List<Tuple<uint, uint>> AliveEntities { get; set; }
             
         // Deserialize용 생성자
         public VerifyStageBattleResponse() { }
